+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0
+
+- Add ApiVersion in RapidOptions class
+- Add ThreeDSecureAuth in RequestPaymentInstrument class
+- Add ThreeDSecureAuth & PaymentType in ResponsePaymentInstrument class
+- Add CardDetails under Customer section for transaction query API's response
+
 ## 0.1.0
 
 - First release